==> Machine Learning/Practice/CampusX/Feature Engineering/11_Feature_Encoding(Ordinal_and_Label_Encoder).ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('customer.csv')

df.sample(5)

df = df.iloc[:,2:]

df.head()

from sklearn.preprocessing import OrdinalEncoder

X_train

oe = OrdinalEncoder(categories=[['Poor','Average','Good'],['School','UG','PG']])

oe.fit(X_train)

X_train = oe.transform(X_train)

X_train

oe.categories_

X_train

from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()

le.fit(y_train)

le.classes_

y_train = le.transform(y_train)
y_test = le.transform(y_test)

y_train
