==> Machine Learning/Practice/CampusX/Feature Engineering/13_Feature_Encoding(Column_Transformer).ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('D:\\Study\\Data-Science\\Machine Learning\\Practice\\CampusX\\Feature Engineering\\Datasets\\covid_toy.csv')

df.sample(10)

We will do:

        OHE -> gender, city
        
        Ordinal encoding -> cough
        
        simple Imputer -> fever
        
        Label encoder -> has_covid

### Train, test, split

from sklearn.model_selection import train_test_split

X = df.iloc[:,:-1]
X

y = df['has_covid']
y

### Manual approach to encode and impute features

df['fever'].isnull().sum()

### Why we can't do encoding and imputing before splitting data? Why we have to do that after splitting!

The reason for performing data splitting before imputing and encoding in a machine learning pipeline is to avoid data leakage and ensure that your model's evaluation is a realistic representation of its performance on unseen data. Let's elaborate on this:

Data Leakage: Data leakage refers to a situation where information from the testing set unintentionally influences the training process. It can lead to overly optimistic model evaluation because the model has "seen" some of the test data during preprocessing. This can result in models that do not generalize well to new, unseen data.

Training and Testing Sets Must Be Independent: The fundamental principle of machine learning is that the model should learn from one set of data (the training set) and be evaluated on a separate, independent set of data (the testing set). This separation ensures that the model's performance assessment is not biased and that it can handle new, unseen data.

Imputing and Encoding Based on the Training Data: When you perform imputation (e.g., filling missing values) and encoding (e.g., one-hot encoding) on the entire dataset before splitting, the preprocessing steps consider information from both the training and testing data. This leads to two problems:

a. Information Leakage: Imputing and encoding with the entire dataset can introduce information from the testing set into the training set, violating the independence principle. For example, if you use the mean for imputation, the mean value computed on the entire dataset will include information from the testing set.

b. Optimistic Evaluation: If you perform imputation and encoding on the entire dataset before splitting, your model might perform well on the test set due to the information it "leaked" during preprocessing. However, this doesn't reflect how well the model generalizes to truly unseen data.

Preventing Overfitting: Data leakage can also lead to overfitting, where the model captures noise in the data instead of the underlying patterns. Proper separation of the training and testing data helps mitigate overfitting.

By splitting the data into a training set and a testing set first and then performing imputing and encoding separately for each set, you ensure that your model is trained on clean, independent data and that your evaluation is free from the influence of the testing set. This approach provides a more accurate representation of your model's real-world performance and ensures that it can generalize well to new, unseen data.

X_train,X_test,y_train,y_test = train_test_split(X,y,test_size=0.2,random_state=0)

X_train.shape

X_test.shape

#### Adding simple imputer to fever columns

from sklearn.impute import SimpleImputer

imputer = SimpleImputer()

X_train_fever = imputer.fit_transform(X_train[['fever']])
X_train_fever.shape

X_test_fever = imputer.transform(X_test[['fever']])
X_test_fever.shape

#### Adding One Hot Encoding to Gender and City columns

from sklearn.preprocessing import OneHotEncoder

X_train[['gender']].value_counts()

X_train[['city']].value_counts()

ohe = OneHotEncoder(drop='first',sparse=False)

X_train_gender_city = ohe.fit_transform(X_train[['gender','city']])

X_train_gender_city.shape

X_test_gender_city = ohe.transform(X_test[['gender','city']])

X_test_gender_city.shape

#### Ordinal Encoding on Cough column

X_train['cough'].value_counts()

from sklearn.preprocessing import OrdinalEncoder

order = ['Mild','Strong']
oe = OrdinalEncoder(categories=[order])

X_train_cough = oe.fit_transform(X_train[['cough']])
X_train_cough.shape

X_test_cough = oe.transform(X_test[['cough']])
X_test_cough.shape

#### Label Encoding on Target column

y_train.value_counts()

from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()

y_train = le.fit_transform(y_train)
y_train.shape

y_test = le.transform(y_test)
y_test.shape

# Extracting Age
X_train_age = X_train.drop(columns=['gender','fever','cough','city']).values

# also the test data
X_test_age = X_test.drop(columns=['gender','fever','cough','city']).values

X_train_age.shape

#### Now we have to add all the imputed and encoded numpy arrays together.

X_train_transformed = np.concatenate((X_train_age,X_train_fever,X_train_gender_city,X_train_cough),axis=1)
# also the test data
X_test_transformed = np.concatenate((X_test_age,X_test_fever,X_test_gender_city,X_test_cough),axis=1)

X_train_transformed.shape

### Column Transformer Technique

from sklearn.compose import ColumnTransformer

transformer = ColumnTransformer(
    transformers=[
        ('tnf1',OneHotEncoder(sparse=False,drop='first'),['gender','city']),
        ('tnf2',OrdinalEncoder(categories=[['Mild','Strong']]),['cough']),
        ('tnf3',SimpleImputer(),['fever'])
    ],
    remainder='passthrough'
)

transformer.fit_transform(X_train).shape

transformer.transform(X_test).shape
